refactor: split feature hotspots and own catalog response DTOs
Break comment read-model, upload service, and static-loader import into
domain modules behind stable barrels, and move academic response schemas
into the catalog feature so OpenAPI imports from features.
