+++ src/features/uploads/server/upload-service.ts
@@ -1,839 +1,32 @@
-import { uploadConfig } from "@/features/uploads/lib/upload-config";
-import { normalizeContentType } from "@/features/uploads/lib/upload-utils";
-import {
-  runUploadSerializableTransaction,
-  UploadError,
-} from "@/features/uploads/server/upload-quota";
-import { Prisma, UploadPendingPhase } from "@/generated/prisma/client";
-import {
-  type AuditLogParams,
-  writeAuditLog,
-} from "@/lib/audit/write-audit-log";
-import { getViewerContext } from "@/lib/auth/viewer-context";
-import { withUserDbContext } from "@/lib/db/prisma";
-import { logAppEvent } from "@/lib/log/app-logger";
-import {
-  deleteStorageObject,
-  headStorageObject,
-} from "@/lib/storage/r2-object";
-import { buildUploadKey } from "@/lib/storage/upload-key";
-
-export const MAX_UPLOAD_EXPIRES_SECONDS = 300;
-const EXPIRED_PENDING_UPLOAD_CLEANUP_BATCH_SIZE = 25;
-const UPLOAD_PUT_LEASE_SECONDS = 120;
-
-export type UploadCreateInput = {
-  contentType: string;
-  filename: string;
-  size: number;
-};
-
-export type UploadCompleteInput = {
-  contentType?: string | null;
-  filename: string;
-  key: string;
-};
-
-type UploadUsagePrisma = {
-  upload: {
-    aggregate: (input: {
-      where: { userId: string };
-      _sum: { size: true };
-    }) => Promise<{ _sum: { size: number | null } }>;
-  };
-  uploadPending: {
-    aggregate: (input: {
-      where: {
-        userId: string;
-        expiresAt: { gt: Date };
-        NOT?: { key: string };
-      };
-      _sum: { size: true };
-    }) => Promise<{ _sum: { size: number | null } }>;
-  };
-};
-
-type ExpiredPendingUploadCleanupPrisma = {
-  uploadPending: {
-    findMany: (input: {
-      where: {
-        userId: string;
-        expiresAt: { lt: Date };
-        phase: { in: UploadPendingPhase[] };
-        OR: Array<{ leaseExpiresAt: null } | { leaseExpiresAt: { lt: Date } }>;
-        NOT?: { key: string };
-      };
-      orderBy: [{ expiresAt: "asc" }, { key: "asc" }];
-      select: { key: true };
-      take: number;
-    }) => Promise<Array<{ key: string }>>;
-    deleteMany: (input: {
-      where: {
-        key: { in: string[] };
-        userId: string;
-        expiresAt: { lt: Date };
-      };
-    }) => Promise<unknown>;
-  };
-};
-
-export const managedUploadSelect = {
-  id: true,
-  key: true,
-  filename: true,
-  size: true,
-  createdAt: true,
-} as const;
-
-type PublicUpload = {
-  createdAt: Date | string;
-  filename: string;
-  id: string;
-  key: string;
-  size: number;
-};
-
-type UploadCompletionFailureCode = "Quota exceeded" | "Upload session expired";
-
-type UploadCompletionResult =
-  | { ok: true; upload: PublicUpload; usedBytes: number }
-  | { ok: false; code: UploadCompletionFailureCode };
-
-type DownloadableUpload = {
-  contentType: string | null;
-  filename: string;
-  key: string;
-  userId: string;
-};
-
-export function publicUploadPayload(upload: PublicUpload) {
-  return {
-    id: upload.id,
-    key: upload.key,
-    filename: upload.filename,
-    size: upload.size,
-    createdAt: upload.createdAt,
-  };
-}
-
-export function uploadKeyBelongsToUser(key: string, userId: string) {
-  return key.startsWith(`uploads/${userId}/`);
-}
-
-export async function createUploadSession(input: {
-  origin: string;
-  upload: UploadCreateInput;
-  userId: string;
-}) {
-  const now = new Date();
-  await withUserDbContext(input.userId, (tx) =>
-    deleteExpiredPendingUploads(tx, input.userId, now),
-  );
-
-  const key = buildUploadKey(input.userId);
-  const expiresAt = new Date(Date.now() + MAX_UPLOAD_EXPIRES_SECONDS * 1000);
-
-  const reservation = await runOwnedUploadSerializableTransaction(
-    input.userId,
-    async (tx) => {
-      const usedBytes = await getUploadUsedBytes({
-        prisma: tx,
-        userId: input.userId,
-        now,
-      });
-      if (usedBytes + input.upload.size > uploadConfig.totalQuotaBytes) {
-        throw new UploadError("Quota exceeded");
-      }
-
-      await tx.uploadPending.create({
-        data: {
-          attemptId: crypto.randomUUID(),
-          contentType: input.upload.contentType,
-          expiresAt,
-          filename: input.upload.filename,
-          key,
-          phase: UploadPendingPhase.reserved,
-          size: input.upload.size,
-          userId: input.userId,
-        },
-      });
-
-      return { usedBytes };
-    },
-    "Failed to reserve upload quota",
-  );
-
-  const uploadUrl = new URL("/api/workspace/uploads/object", input.origin);
-  uploadUrl.searchParams.set("key", key);
-
-  return {
-    key,
-    url: uploadUrl.toString(),
-    maxFileSizeBytes: uploadConfig.maxFileSizeBytes,
-    quotaBytes: uploadConfig.totalQuotaBytes,
-    usedBytes: reservation.usedBytes,
-  };
-}
-
-export async function createOwnedUploadSession(input: {
-  origin: string;
-  upload: UploadCreateInput;
-  userId: string;
-}) {
-  const writer = await requireActiveUploadWriter(input.userId);
-  if (!writer.ok) return writer;
-
-  return {
-    ok: true as const,
-    session: await createUploadSession(input),
-  };
-}
-
-export async function completeUploadSession(
-  userId: string,
-  input: UploadCompleteInput,
-) {
-  const now = new Date();
-  await withUserDbContext(userId, (tx) =>
-    deleteExpiredPendingUploads(tx, userId, now, input.key),
-  );
-  if (!uploadKeyBelongsToUser(input.key, userId)) {
-    throw new UploadError("Upload session expired");
-  }
-
-  const existing = await withUserDbContext(userId, (tx) =>
-    findExistingUploadUsagePayload(tx, input.key, userId, now),
-  );
-  if (existing) return existing;
-
-  try {
-    await withUserDbContext(userId, (tx) =>
-      assertActivePendingUpload(tx, {
-        key: input.key,
-        now,
-        userId,
-      }),
-    );
-  } catch (error) {
-    if (error instanceof UploadError) {
-      const completed = await withUserDbContext(userId, (tx) =>
-        findExistingUploadUsagePayload(tx, input.key, userId, new Date()),
-      );
-      if (completed) return completed;
-    }
-    throw error;
-  }
-
-  const uploadedObject = await validateUploadedObject(input);
-
-  const reservation = await runOwnedUploadSerializableTransaction(
-    userId,
-    async (tx) => {
-      const completed = await tx.upload.findUnique({
-        where: { key: input.key },
-      });
-      if (completed) {
-        if (completed.userId !== userId) {
-          return {
-            ok: false,
-            code: "Upload session expired",
-          } satisfies UploadCompletionResult;
-        }
-        await tx.uploadPending.deleteMany({
-          where: { key: input.key, userId },
-        });
-        const usedBytes = await getUploadUsedBytes({
-          prisma: tx,
-          userId,
-          now: new Date(),
-        });
-        return {
-          ok: true,
-          upload: completed,
-          usedBytes: usedBytes || completed.size,
-        } satisfies UploadCompletionResult;
-      }
-
-      const pending = await tx.uploadPending.findUnique({
-        where: { key: input.key },
-      });
-      if (!pending || pending.userId !== userId) {
-        return {
-          ok: false,
-          code: "Upload session expired",
-        } satisfies UploadCompletionResult;
-      }
-
-      const transactionNow = new Date();
-      if (pending.expiresAt < transactionNow) {
-        return {
-          ok: false,
-          code: "Upload session expired",
-        } satisfies UploadCompletionResult;
-      }
-
-      if (pending.phase !== UploadPendingPhase.uploaded) {
-        return {
-          ok: false,
-          code: "Upload session expired",
-        } satisfies UploadCompletionResult;
-      }
-
-      const completing = await tx.uploadPending.updateMany({
-        where: {
-          key: input.key,
-          userId,
-          phase: UploadPendingPhase.uploaded,
-        },
-        data: {
-          phase: UploadPendingPhase.completing,
-          leaseExpiresAt: new Date(transactionNow.getTime() + 30_000),
-        },
-      });
-      if (completing.count === 0) {
-        return {
-          ok: false,
-          code: "Upload session expired",
-        } satisfies UploadCompletionResult;
-      }
-
-      const usedBytes = await getUploadUsedBytes({
-        excludePendingKey: input.key,
-        prisma: tx,
-        userId,
-        now: transactionNow,
-      });
-      if (usedBytes + uploadedObject.size > uploadConfig.totalQuotaBytes) {
-        await tx.uploadPending.updateMany({
-          where: {
-            key: input.key,
-            userId,
-            expiresAt: { gte: transactionNow },
-          },
-          data: { expiresAt: new Date(transactionNow.getTime() - 1) },
-        });
-        return {
-          ok: false,
-          code: "Quota exceeded",
-        } satisfies UploadCompletionResult;
-      }
-
-      const upload = await tx.upload.create({
-        data: {
-          contentType: uploadedObject.contentType,
-          filename: input.filename,
-          key: input.key,
-          size: uploadedObject.size,
-          userId,
-        },
-      });
-
-      await tx.uploadPending.deleteMany({ where: { key: input.key, userId } });
-
-      return {
-        ok: true,
-        upload,
-        usedBytes: usedBytes + uploadedObject.size,
-      } satisfies UploadCompletionResult;
-    },
-    "Failed to finalize upload quota",
-  );
-
-  if (!reservation.ok) {
-    throw new UploadError(reservation.code);
-  }
-
-  return uploadUsagePayload(reservation.upload, reservation.usedBytes);
-}
-
-export async function completeOwnedUploadSession(
-  userId: string,
-  input: UploadCompleteInput,
-) {
-  const writer = await requireActiveUploadWriter(userId);
-  if (!writer.ok) return writer;
-  if (!uploadKeyBelongsToUser(input.key, userId)) {
-    return { ok: false as const, error: "forbidden" as const };
-  }
-
-  return {
-    ok: true as const,
-    completion: await completeUploadSession(userId, input),
-  };
-}
-
-export async function listUploads(
-  userId: string,
-  pagination: { pageSize: number; skip: number },
-) {
-  const now = new Date();
-  return withUserDbContext(userId, async (tx) => {
-    const [uploads, total, usage, pendingUsage] = await Promise.all([
-      tx.upload.findMany({
-        where: { userId },
-        orderBy: [{ createdAt: "desc" }, { id: "desc" }],
-        skip: pagination.skip,
-        take: pagination.pageSize,
-        select: managedUploadSelect,
-      }),
-      tx.upload.count({ where: { userId } }),
-      tx.upload.aggregate({
-        where: { userId },
-        _sum: { size: true },
-      }),
-      tx.uploadPending.aggregate({
-        where: { userId, expiresAt: { gt: now } },
-        _sum: { size: true },
-      }),
-    ]);
-
-    const usedBytes = (usage._sum.size ?? 0) + (pendingUsage._sum.size ?? 0);
-
-    return {
-      maxFileSizeBytes: uploadConfig.maxFileSizeBytes,
-      quotaBytes: uploadConfig.totalQuotaBytes,
-      total,
-      uploads: uploads.map(publicUploadPayload),
-      usedBytes,
-    };
-  });
-}
-
-export async function renameUpload(input: {
-  filename: string;
-  id: string;
-  userId: string;
-}) {
-  return withUserDbContext(input.userId, async (tx) => {
-    const upload = await tx.upload.findFirst({
-      where: { id: input.id, userId: input.userId },
-      select: { id: true },
-    });
-
-    if (!upload) return null;
-
-    const updated = await tx.upload.update({
-      where: { id: upload.id },
-      data: { filename: input.filename },
-      select: managedUploadSelect,
-    });
-
-    return publicUploadPayload(updated);
-  });
-}
-
-export async function renameOwnedUpload(input: {
-  filename: string;
-  id: string;
-  userId: string;
-}) {
-  const writer = await requireActiveUploadWriter(input.userId);
-  if (!writer.ok) return writer;
-
-  const upload = await renameUpload(input);
-  if (!upload) return { ok: false as const, error: "not_found" as const };
-  return { ok: true as const, upload };
-}
-
-async function findUploadRecordForDeletion(input: {
-  id: string;
-  userId: string;
-}) {
-  return withUserDbContext(input.userId, async (tx) => {
-    const upload = await tx.upload.findFirst({
-      where: { id: input.id, userId: input.userId },
-      select: { id: true, key: true, size: true },
-    });
-
-    if (!upload) return null;
-    return upload;
-  });
-}
-
-export async function deleteOwnedUpload(input: {
-  audit?: UploadAuditContext;
-  id: string;
-  userId: string;
-}) {
-  const writer = await requireActiveUploadWriter(input.userId);
-  if (!writer.ok) return writer;
-
-  const upload = await findUploadRecordForDeletion({
-    id: input.id,
-    userId: input.userId,
-  });
-  if (!upload) return { ok: false as const, error: "not_found" as const };
-
-  const storageDeleted = await deleteUploadStorageObject(upload);
-  if (!storageDeleted) {
-    return {
-      ok: false as const,
-      error: "storage_delete_failed" as const,
-    };
-  }
-
-  const metadataDeleted = await withUserDbContext(input.userId, async (tx) => {
-    const deleted = await tx.upload.deleteMany({
-      where: { id: upload.id, userId: input.userId },
-    });
-    if (deleted.count === 0) return false;
-
-    await writeUploadDeleteAuditLog({
-      audit: input.audit,
-      client: tx,
-      upload,
-      userId: input.userId,
-    });
-    return true;
-  });
-  if (!metadataDeleted)
-    return { ok: false as const, error: "not_found" as const };
-
-  return {
-    ok: true as const,
-    deletedId: upload.id,
-    deletedSize: upload.size,
-  };
-}
-
-export async function findDownloadableUpload(id: string, userId: string) {
-  const viewer = await getViewerContext({ includeAdmin: true, userId });
-  if (!viewer.isAuthenticated) return null;
-
-  const [upload] = await withUserDbContext(userId, (tx) =>
-    tx.$queryRaw<DownloadableUpload[]>(Prisma.sql`
-      SELECT *
-      FROM public.find_downloadable_upload(${id})
-    `),
-  );
-  return upload ?? null;
-}
-
-export async function claimUploadPutLease(input: {
-  key: string;
-  requestContentLength: number;
-  requestContentType: string | null;
-  userId: string;
-}) {
-  const now = new Date();
-  const leaseExpiresAt = new Date(
-    now.getTime() + UPLOAD_PUT_LEASE_SECONDS * 1000,
-  );
-  const claimed = await withUserDbContext(input.userId, async (tx) => {
-    const updated = await tx.uploadPending.updateMany({
-      where: {
-        key: input.key,
-        userId: input.userId,
-        expiresAt: { gt: now },
-        phase: {
-          in: [UploadPendingPhase.reserved, UploadPendingPhase.uploaded],
-        },
-        OR: [{ leaseExpiresAt: null }, { leaseExpiresAt: { lt: now } }],
-      },
-      data: {
-        leaseExpiresAt,
-        phase: UploadPendingPhase.uploading,
-      },
-    });
-    if (updated.count === 0) {
-      return null;
-    }
-
-    return tx.uploadPending.findUnique({
-      where: { key: input.key },
-      select: {
-        attemptId: true,
-        contentType: true,
-        size: true,
-      },
-    });
-  });
-
-  if (!claimed) {
-    throw new UploadError("Upload session expired");
-  }
-  if (input.requestContentLength > claimed.size) {
-    throw new UploadError("File too large");
-  }
-
-  return {
-    attemptId: claimed.attemptId,
-    contentType:
-      normalizeContentType(input.requestContentType) ?? claimed.contentType,
-  };
-}
-
-export async function markUploadPutCompleted(input: {
-  attemptId: string;
-  key: string;
-  userId: string;
-}) {
-  const updated = await withUserDbContext(input.userId, (tx) =>
-    tx.uploadPending.updateMany({
-      where: {
-        attemptId: input.attemptId,
-        key: input.key,
-        phase: UploadPendingPhase.uploading,
-        userId: input.userId,
-      },
-      data: {
-        leaseExpiresAt: null,
-        phase: UploadPendingPhase.uploaded,
-      },
-    }),
-  );
-
-  if (updated.count === 0) {
-    throw new UploadError("Upload session expired");
-  }
-}
-
-export async function validatePendingUploadObject(input: {
-  key: string;
-  requestContentLength: number;
-  requestContentType: string | null;
-  userId: string;
-}) {
-  return claimUploadPutLease(input);
-}
-
-async function requireActiveUploadWriter(userId: string) {
-  const viewer = await getViewerContext({ includeAdmin: true, userId });
-  if (!viewer.isAuthenticated) {
-    return { ok: false as const, error: "forbidden" as const };
-  }
-  if (viewer.isSuspended) {
-    return {
-      ok: false as const,
-      error: "suspended" as const,
-      reason: viewer.suspensionReason,
-    };
-  }
-  return { ok: true as const };
-}
-
-async function deleteUploadStorageObject(upload: {
-  key: string;
-  size: number;
-}) {
-  try {
-    await deleteStorageObject(upload.key);
-    return true;
-  } catch (error) {
-    try {
-      const head = await headStorageObject(upload.key);
-      if (upload.size > 0 && head.size <= 0) {
-        return true;
-      }
-    } catch {
-      // Preserve the upload record if storage state cannot be confirmed.
-    }
-
-    logAppEvent(
-      "error",
-      "R2 object deletion failed; upload record preserved",
-      { source: "upload" },
-      error,
-    );
-    return false;
-  }
-}
-
-async function writeUploadDeleteAuditLog({
-  audit,
-  client,
-  upload,
-  userId,
-}: {
-  audit?: UploadAuditContext;
-  client: NonNullable<Parameters<typeof writeAuditLog>[1]>;
-  upload: { id: string; key: string; size: number };
-  userId: string;
-}) {
-  const { source, ...attribution } = audit ?? {};
-  await writeAuditLog(
-    {
-      action: "upload_delete",
-      ...attribution,
-      userId,
-      subjectUserId: attribution.subjectUserId ?? userId,
-      targetId: upload.id,
-      targetType: "upload",
-      metadata: {
-        size: upload.size,
-        ...(source ? { source } : {}),
-      },
-    },
-    client,
-  );
-}
-
-type UploadAuditContext = Pick<
-  AuditLogParams,
-  | "channel"
-  | "ipAddress"
-  | "oauthClientId"
-  | "oauthGrantId"
-  | "requestId"
-  | "sessionId"
-  | "subjectUserId"
-  | "userAgent"
-> & { source?: "graphql" | "mcp" };
-
-async function deleteExpiredPendingUploads(
-  uploadPrisma: ExpiredPendingUploadCleanupPrisma,
-  userId: string,
-  now: Date,
-  excludeKey?: string,
-) {
-  const expiredUploads = await uploadPrisma.uploadPending.findMany({
-    where: {
-      userId,
-      expiresAt: { lt: now },
-      phase: {
-        in: [UploadPendingPhase.reserved, UploadPendingPhase.uploaded],
-      },
-      OR: [{ leaseExpiresAt: null }, { leaseExpiresAt: { lt: now } }],
-      ...(excludeKey ? { NOT: { key: excludeKey } } : {}),
-    },
-    orderBy: [{ expiresAt: "asc" }, { key: "asc" }],
-    select: { key: true },
-    take: EXPIRED_PENDING_UPLOAD_CLEANUP_BATCH_SIZE,
-  });
-  if (expiredUploads.length === 0) return;
-
-  await uploadPrisma.uploadPending.deleteMany({
-    where: {
-      key: { in: expiredUploads.map(({ key }) => key) },
-      userId,
-      expiresAt: { lt: now },
-    },
-  });
-}
-
-async function deletePendingUpload(
-  uploadPrisma: Prisma.TransactionClient,
-  key: string,
-  userId: string,
-) {
-  await uploadPrisma.uploadPending.deleteMany({
-    where: { key, userId },
-  });
-}
-
-async function findExistingUploadUsagePayload(
-  uploadPrisma: Prisma.TransactionClient,
-  key: string,
-  userId: string,
-  now: Date,
-) {
-  const existing = await uploadPrisma.upload.findUnique({
-    where: { key },
-  });
-  if (!existing) return null;
-  if (existing.userId !== userId) {
-    throw new UploadError("Upload session expired");
-  }
-
-  await deletePendingUpload(uploadPrisma, key, userId);
-  const usedBytes = await getUploadUsedBytes({
-    prisma: uploadPrisma,
-    userId,
-    now,
-  });
-  return uploadUsagePayload(existing, usedBytes || existing.size);
-}
-
-async function assertActivePendingUpload(
-  uploadPrisma: Prisma.TransactionClient,
-  input: {
-    key: string;
-    now: Date;
-    userId: string;
-  },
-) {
-  const pending = await uploadPrisma.uploadPending.findUnique({
-    where: { key: input.key },
-    select: {
-      expiresAt: true,
-      userId: true,
-    },
-  });
-
-  if (!pending || pending.userId !== input.userId) {
-    throw new UploadError("Upload session expired");
-  }
-
-  if (pending.expiresAt < input.now) {
-    throw new UploadError("Upload session expired");
-  }
-}
-
-async function runOwnedUploadSerializableTransaction<T>(
-  userId: string,
-  action: (tx: Prisma.TransactionClient) => Promise<T>,
-  failureMessage: string,
-) {
-  const normalizedUserId = userId.trim();
-  if (!normalizedUserId) throw new Error("RLS user ID is required");
-
-  return runUploadSerializableTransaction(async (tx) => {
-    await tx.$queryRaw`SELECT set_config('app.user_id', ${normalizedUserId}, true)`;
-    return action(tx);
-  }, failureMessage);
-}
-
-async function getUploadUsedBytes(input: {
-  excludePendingKey?: string;
-  now: Date;
-  prisma: UploadUsagePrisma;
-  userId: string;
-}) {
-  const [usage, pendingUsage] = await Promise.all([
-    input.prisma.upload.aggregate({
-      where: { userId: input.userId },
-      _sum: { size: true },
-    }),
-    input.prisma.uploadPending.aggregate({
-      where: {
-        userId: input.userId,
-        expiresAt: { gt: input.now },
-        ...(input.excludePendingKey
-          ? { NOT: { key: input.excludePendingKey } }
-          : {}),
-      },
-      _sum: { size: true },
-    }),
-  ]);
-
-  return (usage._sum.size ?? 0) + (pendingUsage._sum.size ?? 0);
-}
-
-async function validateUploadedObject(input: {
-  contentType?: string | null;
-  key: string;
-}) {
-  const head = await headStorageObject(input.key);
-
-  const size = head.size;
-  if (!size || size <= 0) {
-    throw new UploadError("Uploaded object missing");
-  }
-
-  if (size > uploadConfig.maxFileSizeBytes) {
-    throw new UploadError("File too large");
-  }
-
-  return {
-    contentType: normalizeContentType(input.contentType) ?? head.contentType,
-    size,
-  };
-}
-
-function uploadUsagePayload(upload: PublicUpload, usedBytes: number) {
-  return {
-    upload: publicUploadPayload(upload),
-    usedBytes,
-    quotaBytes: uploadConfig.totalQuotaBytes,
-  };
-}
+/** Upload service barrel — re-exports split domain modules. */
+
+export {
+  completeOwnedUploadSession,
+  completeUploadSession,
+  type UploadCompleteInput,
+} from "./upload-complete-session";
+export {
+  createOwnedUploadSession,
+  createUploadSession,
+  type UploadCreateInput,
+} from "./upload-create-session";
+export {
+  listUploads,
+  publicUploadPayload,
+  uploadKeyBelongsToUser,
+} from "./upload-list";
+export {
+  deleteOwnedUpload,
+  findDownloadableUpload,
+  renameOwnedUpload,
+  renameUpload,
+} from "./upload-manage";
+export {
+  claimUploadPutLease,
+  markUploadPutCompleted,
+  validatePendingUploadObject,
+} from "./upload-put-lease";
+export {
+  MAX_UPLOAD_EXPIRES_SECONDS,
+  managedUploadSelect,
+} from "./upload-service-shared";
