+++ src/static-loader/import.ts
@@ -1,3 +1,4 @@
+/** Static snapshot import orchestration (transaction boundaries stay here). */
 import type { Prisma, PrismaClient } from "../generated/prisma-node/client";
 import {
   loadCatalogLookups,
@@ -6,46 +7,56 @@ import {
   loadScheduleLookups,
   loadSemesters,
 } from "./catalog-plan";
-import {
-  bulkUpsert,
-  type ColumnValue,
-  deleteMissingSnapshotRows,
-} from "./database-writes";
+import { deleteMissingSnapshotRows } from "./database-writes";
 import { collectCodeOnlyDepartmentPlaceholders } from "./department-placeholder";
+import {
+  upsertCourses,
+  upsertExamBatches,
+  upsertTeacherLessonTypes,
+  upsertTeachers,
+  upsertTeacherTitles,
+} from "./import-courses";
+import {
+  upsertAdminClasses,
+  upsertBuildings,
+  upsertCampuses,
+  upsertRooms,
+  upsertRoomTypes,
+} from "./import-infrastructure";
 import { acquireStaticImportLock } from "./import-lock";
+import {
+  loadLookupTables,
+  upsertDepartments,
+  upsertSemesters,
+} from "./import-lookups";
+import {
+  upsertExams,
+  upsertScheduleGroups,
+  upsertSections,
+  writeExamRooms,
+} from "./import-sections";
 import {
   assertStaticImportStateAllowsSnapshot,
   recordStaticImportState,
   STATIC_IMPORT_TRANSFORM_REVISION,
 } from "./import-state";
+import type {
+  ImportConfig,
+  ImportRecordCounts,
+  ImportReport,
+} from "./import-types";
+import { countStats, syncYoungEvents } from "./import-young";
 import { loadScheduleInfrastructure } from "./infrastructure-plan";
 import type {
-  AdminClassBuild,
   AdminClassSectionPair,
-  BuildingBuild,
-  CampusBuild,
-  CourseBuild,
-  DepartmentBuild,
-  DepartmentPlaceholderRequest,
-  ExamBatchBuild,
-  ExamBuild,
-  RoomBuild,
-  RoomTypeBuild,
-  ScheduleGroupBuild,
-  SectionBuild,
   SectionTeacherPair,
-  SemesterBuild,
   TeacherAssignmentBuild,
-  TeacherBuild,
-  TeacherLessonTypeBuild,
-  TeacherTitleBuild,
 } from "./mappers";
 import {
   writeAdminClassSections,
   writeSectionTeachers,
   writeTeacherAssignments,
 } from "./relation-writes";
-import { optionalId, requiredId } from "./required-id";
 import { writeSchedules } from "./schedule-writes";
 import {
   reconcileSectionPresence,
@@ -59,45 +70,14 @@ import {
   validateMappedSectionJwIds,
   validateSnapshotCompleteness,
 } from "./validation";
-import { loadYoungEvents, type YoungEventBuild } from "./young-plan";
+import { loadYoungEvents } from "./young-plan";
 
-export type ImportConfig = {
-  snapshotPath: string;
-  snapshotSha256: string;
-  minSemester: number;
-  dryRun: boolean;
-};
-
-export type ImportRecordCounts = {
-  semesters: number;
-  departments: number;
-  courses: number;
-  sections: number;
-  teachers: number;
-  scheduleGroups: number;
-  schedules: number;
-  exams: number;
-  rooms: number;
-  buildings: number;
-  campuses: number;
-  adminClasses: number;
-  youngEvents: number;
-};
-
-export type ImportReport = {
-  mode: "apply" | "dry-run";
-  outcome: "committed" | "rolled-back" | "unchanged";
-  snapshot: {
-    sha256: string;
-    schemaVersion: string;
-    generatedAt: string | null;
-  };
-  plannedRecordCounts: ImportRecordCounts | null;
-  databaseRecordCounts: ImportRecordCounts | null;
-  reconciliation: {
-    sectionPresence: SectionPresenceStats | { status: "already-applied" };
-  };
-};
+export { upsertAdminClasses } from "./import-infrastructure";
+export type {
+  ImportConfig,
+  ImportRecordCounts,
+  ImportReport,
+} from "./import-types";
 
 export async function runImport(
   prisma: PrismaClient,
@@ -496,957 +476,5 @@ export async function runImport(
     reconciliation: {
       sectionPresence: sectionPresenceStats,
     },
-  };
-}
-
-async function upsertSemesters(
-  tx: Prisma.TransactionClient,
-  builds: SemesterBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "Semester",
-    "jwId",
-    "int",
-    ["nameCn", "code", "startDate", "endDate"],
-    ["text", "text", "date", "date"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [build.nameCn, build.code, build.start, build.end],
-    })),
-  );
-}
-
-async function upsertDepartments(
-  tx: Prisma.TransactionClient,
-  builds: DepartmentBuild[],
-  placeholders: DepartmentPlaceholderRequest[],
-): Promise<Map<string, number>> {
-  const jwIdToId = new Map<number, number>();
-  for (const build of builds) {
-    const result = await tx.department.upsert({
-      where: { jwId: build.jwId },
-      create: {
-        jwId: build.jwId,
-        code: build.code,
-        nameCn: build.nameCn,
-        nameEn: build.nameEn,
-        isCollege: build.isCollege,
-      },
-      update: {
-        code: build.code,
-        nameCn: build.nameCn,
-        nameEn: build.nameEn,
-        isCollege: build.isCollege,
-      },
-    });
-    jwIdToId.set(build.jwId, result.id);
-  }
-  const map = new Map<string, number>();
-  for (const build of builds) {
-    const id = jwIdToId.get(build.jwId);
-    if (id != null) map.set(build.code, id);
-  }
-  for (const placeholder of placeholders) {
-    const existing = await tx.department.findFirst({
-      where: { code: placeholder.code, jwId: null },
-      select: { id: true },
-    });
-    const id =
-      existing?.id ??
-      (
-        await tx.department.create({
-          data: {
-            jwId: null,
-            code: placeholder.code,
-            nameCn: placeholder.nameCn,
-            isCollege: false,
-          },
-          select: { id: true },
-        })
-      ).id;
-    map.set(placeholder.code, id);
-  }
-  return map;
-}
-
-type LookupMaps = {
-  courseCategory: Map<string, number>;
-  courseClassify: Map<string, number>;
-  courseGradation: Map<string, number>;
-  courseType: Map<string, number>;
-  educationLevel: Map<string, number>;
-  classType: Map<string, number>;
-  examMode: Map<string, number>;
-  teachLanguage: Map<string, number>;
-};
-
-async function loadLookupTables(
-  tx: Prisma.TransactionClient,
-  lookups: {
-    courseCategories: { nameCn: string; nameEn?: string }[];
-    courseClassifies: { nameCn: string; nameEn?: string }[];
-    courseGradations: { nameCn: string; nameEn?: string }[];
-    courseTypes: { nameCn: string; nameEn?: string }[];
-    educationLevels: { nameCn: string; nameEn?: string }[];
-    classTypes: { nameCn: string; nameEn?: string }[];
-    examModes: { nameCn: string; nameEn?: string }[];
-    teachLanguages: { nameCn: string; nameEn?: string }[];
-  },
-): Promise<LookupMaps> {
-  async function loadModel(
-    table:
-      | "CourseCategory"
-      | "CourseClassify"
-      | "CourseGradation"
-      | "CourseType"
-      | "EducationLevel"
-      | "ClassType"
-      | "ExamMode"
-      | "TeachLanguage",
-    items: { nameCn: string; nameEn?: string }[],
-  ) {
-    return bulkUpsert(
-      tx,
-      table,
-      "nameCn",
-      "text",
-      ["nameEn"],
-      ["text"],
-      items.map((item) => ({ key: item.nameCn, values: [item.nameEn] })),
-    );
-  }
-
-  return {
-    courseCategory: await loadModel("CourseCategory", lookups.courseCategories),
-    courseClassify: await loadModel("CourseClassify", lookups.courseClassifies),
-    courseGradation: await loadModel(
-      "CourseGradation",
-      lookups.courseGradations,
-    ),
-    courseType: await loadModel("CourseType", lookups.courseTypes),
-    educationLevel: await loadModel("EducationLevel", lookups.educationLevels),
-    classType: await loadModel("ClassType", lookups.classTypes),
-    examMode: await loadModel("ExamMode", lookups.examModes),
-    teachLanguage: await loadModel("TeachLanguage", lookups.teachLanguages),
-  };
-}
-
-async function upsertCourses(
-  tx: Prisma.TransactionClient,
-  builds: CourseBuild[],
-  lookupMaps: LookupMaps,
-): Promise<Map<number, number>> {
-  const columns = [
-    "code",
-    "nameCn",
-    "nameEn",
-    "categoryId",
-    "classTypeId",
-    "classifyId",
-    "educationLevelId",
-    "gradationId",
-    "typeId",
-  ];
-  const records = builds.map((build) => ({
-    key: build.jwId,
-    values: [
-      build.code,
-      build.nameCn,
-      build.nameEn,
-      optionalId(
-        lookupMaps.courseCategory,
-        build.categoryName,
-        `CourseCategory ${build.categoryName} for Course jwId ${build.jwId}`,
-      ),
-      optionalId(
-        lookupMaps.classType,
-        build.classTypeName,
-        `ClassType ${build.classTypeName} for Course jwId ${build.jwId}`,
-      ),
-      optionalId(
-        lookupMaps.courseClassify,
-        build.classifyName,
-        `CourseClassify ${build.classifyName} for Course jwId ${build.jwId}`,
-      ),
-      optionalId(
-        lookupMaps.educationLevel,
-        build.educationLevelName,
-        `EducationLevel ${build.educationLevelName} for Course jwId ${build.jwId}`,
-      ),
-      optionalId(
-        lookupMaps.courseGradation,
-        build.gradationName,
-        `CourseGradation ${build.gradationName} for Course jwId ${build.jwId}`,
-      ),
-      optionalId(
-        lookupMaps.courseType,
-        build.typeName,
-        `CourseType ${build.typeName} for Course jwId ${build.jwId}`,
-      ),
-    ] satisfies ColumnValue[],
-  }));
-  return bulkUpsert(
-    tx,
-    "Course",
-    "jwId",
-    "int",
-    columns,
-    ["text", "text", "text", "int", "int", "int", "int", "int", "int"],
-    records,
-  );
-}
-
-async function upsertTeacherTitles(
-  tx: Prisma.TransactionClient,
-  builds: TeacherTitleBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "TeacherTitle",
-    "jwId",
-    "int",
-    ["nameCn", "nameEn", "code", "enabled"],
-    ["text", "text", "text", "boolean"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [build.nameCn, build.nameEn, build.code, build.enabled],
-    })),
-  );
-}
-
-async function upsertTeacherLessonTypes(
-  tx: Prisma.TransactionClient,
-  builds: TeacherLessonTypeBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "TeacherLessonType",
-    "jwId",
-    "int",
-    ["nameCn", "nameEn", "code", "role", "enabled"],
-    ["text", "text", "text", "text", "boolean"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [
-        build.nameCn,
-        build.nameEn,
-        build.code,
-        build.role,
-        build.enabled,
-      ],
-    })),
-  );
-}
-
-async function upsertExamBatches(
-  tx: Prisma.TransactionClient,
-  builds: ExamBatchBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "ExamBatch",
-    "jwId",
-    "int",
-    ["nameCn"],
-    ["text"],
-    builds.map((build) => ({ key: build.jwId, values: [build.nameCn] })),
-  );
-}
-
-async function upsertTeachers(
-  tx: Prisma.TransactionClient,
-  builds: TeacherBuild[],
-  departmentMap: Map<string, number>,
-): Promise<TeacherMap> {
-  const columns = [
-    "personId",
-    "code",
-    "nameCn",
-    "nameEn",
-    "age",
-    "email",
-    "telephone",
-    "mobile",
-    "address",
-    "postcode",
-    "qq",
-    "wechat",
-    "departmentId",
-  ];
-
-  const resolved = builds.map((build) => {
-    const departmentId = build.departmentCode
-      ? departmentMap.get(build.departmentCode)
-      : undefined;
-    return {
-      build,
-      departmentId,
-      key: build.jwId,
-      values: [
-        build.personId ?? null,
-        build.code ?? null,
-        build.nameCn,
-        build.nameEn ?? null,
-        build.age ?? null,
-        build.email ?? null,
-        build.telephone ?? null,
-        build.mobile ?? null,
-        build.address ?? null,
-        build.postcode ?? null,
-        build.qq ?? null,
-        build.wechat ?? null,
-        departmentId ?? null,
-      ] satisfies ColumnValue[],
-    };
-  });
-  const unresolvedTeacherDepartmentCodes = new Set(
-    resolved
-      .filter(
-        ({ build, departmentId }) =>
-          build.departmentCode && departmentId == null,
-      )
-      .map(({ build }) => build.departmentCode as string),
-  );
-  if (unresolvedTeacherDepartmentCodes.size > 0) {
-    throw new Error(
-      `Teacher department codes have no authoritative upstream Department id: ${[...unresolvedTeacherDepartmentCodes].sort().join(", ")}`,
-    );
-  }
-  return bulkUpsert(
-    tx,
-    "Teacher",
-    "jwId",
-    "int",
-    columns,
-    [
-      "int",
-      "text",
-      "text",
-      "text",
-      "int",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "int",
-    ],
-    resolved,
-  );
-}
-
-type TeacherMap = Map<number, number>;
-
-async function upsertCampuses(
-  tx: Prisma.TransactionClient,
-  builds: CampusBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "Campus",
-    "jwId",
-    "int",
-    ["nameCn", "nameEn", "code"],
-    ["text", "text", "text"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [build.nameCn, build.nameEn, build.code],
-    })),
-  );
-}
-
-async function upsertRoomTypes(
-  tx: Prisma.TransactionClient,
-  builds: RoomTypeBuild[],
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "RoomType",
-    "jwId",
-    "int",
-    ["nameCn", "nameEn", "code"],
-    ["text", "text", "text"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [build.nameCn, build.nameEn, build.code],
-    })),
-  );
-}
-
-async function upsertBuildings(
-  tx: Prisma.TransactionClient,
-  builds: BuildingBuild[],
-  campusMap: Map<number, number>,
-): Promise<Map<number, number>> {
-  return bulkUpsert(
-    tx,
-    "Building",
-    "jwId",
-    "int",
-    ["nameCn", "nameEn", "code", "campusId"],
-    ["text", "text", "text", "int"],
-    builds.map((build) => ({
-      key: build.jwId,
-      values: [
-        build.nameCn,
-        build.nameEn,
-        build.code,
-        build.campusJwId == null
-          ? null
-          : requiredId(
-              campusMap,
-              build.campusJwId,
-              `Campus jwId ${build.campusJwId} for Building jwId ${build.jwId}`,
-            ),
-      ],
-    })),
-  );
-}
-
-async function upsertRooms(
-  tx: Prisma.TransactionClient,
-  builds: RoomBuild[],
-  buildingMap: Map<number, number>,
-  roomTypeMap: Map<number, number>,
-): Promise<Map<number, number>> {
-  const columns = [
-    "nameCn",
-    "nameEn",
-    "code",
-    "floor",
-    "virtual",
-    "seatsForSection",
-    "remark",
-    "seats",
-    "buildingId",
-    "roomTypeId",
-  ];
-  const records = builds.map((build) => ({
-    key: build.jwId,
-    values: [
-      build.nameCn,
-      build.nameEn,
-      build.code,
-      build.floor,
-      build.virtual,
-      build.seatsForSection,
-      build.remark,
-      build.seats,
-      build.buildingJwId == null
-        ? null
-        : requiredId(
-            buildingMap,
-            build.buildingJwId,
-            `Building jwId ${build.buildingJwId} for Room jwId ${build.jwId}`,
-          ),
-      build.roomTypeJwId == null
-        ? null
-        : requiredId(
-            roomTypeMap,
-            build.roomTypeJwId,
-            `RoomType jwId ${build.roomTypeJwId} for Room jwId ${build.jwId}`,
-          ),
-    ] satisfies ColumnValue[],
-  }));
-  return bulkUpsert(
-    tx,
-    "Room",
-    "jwId",
-    "int",
-    columns,
-    [
-      "text",
-      "text",
-      "text",
-      "int",
-      "boolean",
-      "int",
-      "text",
-      "int",
-      "int",
-      "int",
-    ],
-    records,
-  );
-}
-
-export async function upsertAdminClasses(
-  tx: Prisma.TransactionClient,
-  builds: AdminClassBuild[],
-): Promise<Map<number, number>> {
-  const columns = [
-    "code",
-    "grade",
-    "nameCn",
-    "nameEn",
-    "stdCount",
-    "planCount",
-    "enabled",
-    "abbrZh",
-    "abbrEn",
-  ];
-  const records = builds.map((build) => ({
-    key: build.jwId,
-    values: [
-      build.code,
-      build.grade,
-      build.nameCn,
-      build.nameEn,
-      build.stdCount,
-      build.planCount,
-      build.enabled,
-      build.abbrZh,
-      build.abbrEn,
-    ] satisfies ColumnValue[],
-  }));
-  return bulkUpsert(
-    tx,
-    "AdminClass",
-    "jwId",
-    "int",
-    columns,
-    ["text", "text", "text", "text", "int", "int", "boolean", "text", "text"],
-    records,
-  );
-}
-
-async function upsertSections(
-  tx: Prisma.TransactionClient,
-  builds: SectionBuild[],
-  semesterMap: Map<number, number>,
-  departmentMap: Map<string, number>,
-  courseMap: Map<number, number>,
-  lookupMaps: LookupMaps,
-  campusMap: Map<number, number>,
-  roomTypeMap: Map<number, number>,
-): Promise<Map<number, number>> {
-  const columns = [
-    "code",
-    "bizTypeId",
-    "credits",
-    "period",
-    "periodsPerWeek",
-    "timesPerWeek",
-    "stdCount",
-    "limitCount",
-    "graduateAndPostgraduate",
-    "dateTimePlaceText",
-    "dateTimePlacePersonText",
-    "actualPeriods",
-    "theoryPeriods",
-    "practicePeriods",
-    "experimentPeriods",
-    "machinePeriods",
-    "designPeriods",
-    "testPeriods",
-    "scheduleState",
-    "suggestScheduleWeeks",
-    "suggestScheduleWeekInfo",
-    "scheduleJsonParams",
-    "selectedStdCount",
-    "remark",
-    "scheduleRemark",
-    "courseId",
-    "semesterId",
-    "campusId",
-    "examModeId",
-    "openDepartmentId",
-    "teachLanguageId",
-    "roomTypeId",
-  ];
-  const records: Array<{ key: number; values: ColumnValue[] }> = [];
-  const unresolvedDepartmentCodes = new Set<string>();
-  for (const build of builds) {
-    const courseId = courseMap.get(build.courseJwId);
-    if (courseId == null) {
-      throw new Error(
-        `Course jwId ${build.courseJwId} did not resolve for section jwId ${build.jwId}`,
-      );
-    }
-    const semesterId = requiredId(
-      semesterMap,
-      build.semesterCode,
-      `Semester jwId ${build.semesterCode} for Section jwId ${build.jwId}`,
-    );
-    const openDepartmentId = build.openDepartmentCode
-      ? departmentMap.get(build.openDepartmentCode)
-      : null;
-    if (build.openDepartmentCode && openDepartmentId == null) {
-      unresolvedDepartmentCodes.add(build.openDepartmentCode);
-    }
-    const campusId =
-      build.campusId == null ? null : campusMap.get(build.campusId);
-    if (campusId == null && build.campusId != null) {
-      throw new Error(
-        `Campus jwId ${build.campusId} did not resolve for section jwId ${build.jwId}`,
-      );
-    }
-    records.push({
-      key: build.jwId,
-      values: [
-        build.code,
-        build.bizTypeId,
-        build.credits,
-        build.period,
-        build.periodsPerWeek,
-        build.timesPerWeek,
-        build.stdCount,
-        build.limitCount,
-        build.graduateAndPostgraduate,
-        build.dateTimePlaceText,
-        build.dateTimePlacePersonText != null
-          ? JSON.stringify(build.dateTimePlacePersonText)
-          : null,
-        build.actualPeriods,
-        build.theoryPeriods,
-        build.practicePeriods,
-        build.experimentPeriods,
-        build.machinePeriods,
-        build.designPeriods,
-        build.testPeriods,
-        build.scheduleState,
-        build.suggestScheduleWeeks != null
-          ? JSON.stringify(build.suggestScheduleWeeks)
-          : null,
-        build.suggestScheduleWeekInfo,
-        build.scheduleJsonParams != null
-          ? JSON.stringify(build.scheduleJsonParams)
-          : null,
-        build.selectedStdCount,
-        build.remark,
-        build.scheduleRemark,
-        courseId,
-        semesterId,
-        campusId,
-        build.examModeName == null
-          ? null
-          : requiredId(
-              lookupMaps.examMode,
-              build.examModeName,
-              `ExamMode ${build.examModeName} for Section jwId ${build.jwId}`,
-            ),
-        openDepartmentId,
-        build.teachLanguageName == null
-          ? null
-          : requiredId(
-              lookupMaps.teachLanguage,
-              build.teachLanguageName,
-              `TeachLanguage ${build.teachLanguageName} for Section jwId ${build.jwId}`,
-            ),
-        build.roomTypeId == null
-          ? null
-          : requiredId(
-              roomTypeMap,
-              build.roomTypeId,
-              `RoomType jwId ${build.roomTypeId} for Section jwId ${build.jwId}`,
-            ),
-      ],
-    });
-  }
-  if (unresolvedDepartmentCodes.size > 0) {
-    throw new Error(
-      `Section department codes have no authoritative upstream Department id: ${[...unresolvedDepartmentCodes].sort().join(", ")}`,
-    );
-  }
-  return bulkUpsert(
-    tx,
-    "Section",
-    "jwId",
-    "int",
-    columns,
-    [
-      "text",
-      "int",
-      "float8",
-      "int",
-      "float8",
-      "int",
-      "int",
-      "int",
-      "boolean",
-      "text",
-      "jsonb",
-      "float8",
-      "float8",
-      "float8",
-      "float8",
-      "float8",
-      "float8",
-      "float8",
-      "text",
-      "jsonb",
-      "text",
-      "jsonb",
-      "int",
-      "text",
-      "text",
-      "int",
-      "int",
-      "int",
-      "int",
-      "int",
-      "int",
-      "int",
-    ],
-    records,
-  );
-}
-
-async function upsertScheduleGroups(
-  tx: Prisma.TransactionClient,
-  builds: ScheduleGroupBuild[],
-  sectionMap: Map<number, number>,
-): Promise<Map<number, number>> {
-  const columns = [
-    "no",
-    "limitCount",
-    "stdCount",
-    "actualPeriods",
-    "isDefault",
-    "sectionId",
-  ];
-  const records: Array<{ key: number; values: ColumnValue[] }> = [];
-  for (const build of builds) {
-    const sectionId = requiredId(
-      sectionMap,
-      build.lessonJwId,
-      `Section jwId ${build.lessonJwId} for ScheduleGroup jwId ${build.jwId}`,
-    );
-    records.push({
-      key: build.jwId,
-      values: [
-        build.no,
-        build.limitCount,
-        build.stdCount,
-        build.actualPeriods,
-        build.isDefault,
-        sectionId,
-      ],
-    });
-  }
-  return bulkUpsert(
-    tx,
-    "ScheduleGroup",
-    "jwId",
-    "int",
-    columns,
-    ["int", "int", "int", "float8", "boolean", "int"],
-    records,
-  );
-}
-
-async function upsertExams(
-  tx: Prisma.TransactionClient,
-  builds: ExamBuild[],
-  sectionMap: Map<number, number>,
-  examBatchMap: Map<number, number>,
-): Promise<Map<number, number>> {
-  const columns = [
-    "examType",
-    "startTime",
-    "endTime",
-    "examDate",
-    "examTakeCount",
-    "examMode",
-    "examBatchId",
-    "sectionId",
-  ];
-  const records: Array<{ key: number; values: ColumnValue[] }> = [];
-  for (const build of builds) {
-    const sectionId = requiredId(
-      sectionMap,
-      build.sectionJwId,
-      `Section jwId ${build.sectionJwId} for Exam jwId ${build.jwId}`,
-    );
-    const examBatchId =
-      build.examBatchJwId == null
-        ? null
-        : requiredId(
-            examBatchMap,
-            build.examBatchJwId,
-            `ExamBatch jwId ${build.examBatchJwId} for Exam jwId ${build.jwId}`,
-          );
-    records.push({
-      key: build.jwId,
-      values: [
-        build.examType,
-        build.startTime,
-        build.endTime,
-        build.examDate,
-        build.examTakeCount,
-        build.examMode,
-        examBatchId,
-        sectionId,
-      ],
-    });
-  }
-  return bulkUpsert(
-    tx,
-    "Exam",
-    "jwId",
-    "int",
-    columns,
-    ["int", "int", "int", "date", "int", "text", "int", "int"],
-    records,
-  );
-}
-
-async function writeExamRooms(
-  tx: Prisma.TransactionClient,
-  builds: ExamBuild[],
-  examMap: Map<number, number>,
-): Promise<void> {
-  const examIds = builds.map((build) =>
-    requiredId(examMap, build.jwId, `Exam jwId ${build.jwId}`),
-  );
-  if (examIds.length === 0) return;
-  await tx.examRoom.deleteMany({ where: { examId: { in: examIds } } });
-
-  const data: Array<{ examId: number; room: string; count: number }> = [];
-  for (const build of builds) {
-    const examId = requiredId(examMap, build.jwId, `Exam jwId ${build.jwId}`);
-    for (const room of build.rooms) {
-      data.push({ examId, room: room.room, count: room.count });
-    }
-  }
-
-  if (data.length > 0) {
-    await tx.examRoom.createMany({ data });
-  }
-}
-
-async function syncYoungEvents(
-  tx: Prisma.TransactionClient,
-  builds: YoungEventBuild[],
-): Promise<void> {
-  const columns = [
-    "name",
-    "category",
-    "department",
-    "organizer",
-    "status",
-    "registrationStatus",
-    "location",
-    "imageUrl",
-    "hours",
-    "capacity",
-    "appliedCount",
-    "startAt",
-    "endAt",
-    "applyStartAt",
-    "applyEndAt",
-    "isActive",
-    "rawJson",
-  ];
-  await bulkUpsert(
-    tx,
-    "YoungEvent",
-    "youngId",
-    "text",
-    columns,
-    [
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "text",
-      "float8",
-      "int",
-      "int",
-      "timestamp",
-      "timestamp",
-      "timestamp",
-      "timestamp",
-      "boolean",
-      "jsonb",
-    ],
-    builds.map((build) => ({
-      key: build.youngId,
-      values: [
-        build.name,
-        build.category,
-        build.department,
-        build.organizer,
-        build.status,
-        build.registrationStatus,
-        build.location,
-        build.imageUrl,
-        build.hours,
-        build.capacity,
-        build.appliedCount,
-        build.startAt,
-        build.endAt,
-        build.applyStartAt,
-        build.applyEndAt,
-        build.isActive,
-        build.rawJson,
-      ] satisfies ColumnValue[],
-    })),
-  );
-
-  // The snapshot is authoritative for both lists; drop events that disappeared.
-  // An empty snapshot means the upstream fetch broke (the ended list alone
-  // carries thousands of historical events), so keep existing rows instead of
-  // wiping the table.
-  if (builds.length === 0) return;
-  const keepYoungIds = builds.map((build) => build.youngId);
-  await tx.youngEvent.deleteMany({
-    where: { youngId: { notIn: keepYoungIds } },
-  });
-}
-
-async function countStats(
-  prisma: Prisma.TransactionClient,
-): Promise<ImportRecordCounts> {
-  const [
-    semesters,
-    departments,
-    courses,
-    sections,
-    teachers,
-    scheduleGroups,
-    schedules,
-    exams,
-    rooms,
-    buildings,
-    campuses,
-    adminClasses,
-    youngEvents,
-  ] = await Promise.all([
-    prisma.semester.count(),
-    prisma.department.count(),
-    prisma.course.count(),
-    prisma.section.count(),
-    prisma.teacher.count(),
-    prisma.scheduleGroup.count(),
-    prisma.schedule.count(),
-    prisma.exam.count(),
-    prisma.room.count(),
-    prisma.building.count(),
-    prisma.campus.count(),
-    prisma.adminClass.count(),
-    prisma.youngEvent.count(),
-  ]);
-
-  return {
-    semesters,
-    departments,
-    courses,
-    sections,
-    teachers,
-    scheduleGroups,
-    schedules,
-    exams,
-    rooms,
-    buildings,
-    campuses,
-    adminClasses,
-    youngEvents,
   };
 }
